requirements.txt
pandas
sqlparse
psycopg2-binary
sqlalchemy